ec2-api/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
pydantic-settings>=2.0.0
uvicorn[standard]>=0.30.0
